Unify option value retrieval.
