+++ src/CommandLineUtils.Extensions/Options/MappingExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using CommandLineUtils.Extensions.Utilities;
 using McMaster.Extensions.CommandLineUtils;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
@@ -29,21 +30,47 @@ namespace CommandLineUtils.Extensions.Options
                              (o, p) => new { Option = o, Property = p });
 
             foreach (var p in properties)
-                p.Property.SetValue(instance, GetValue(p.Option));
+                p.Property.SetValue(instance, p.Option.GetValue(p.Property.PropertyType));
 
             return instance;
         }
 
-        private static object GetValue(CommandOption option)
+        /// <summary>
+        /// Attempts to get the value of a <see cref="CommandOption"/> given the target type <typeparamref name="T"/>.
+        /// </summary>
+        /// <typeparam name="T">The expected type of the option's value.</typeparam>
+        public static T GetValue<T>(this CommandOption option) => (T)option.GetValue(typeof(T));
+
+        /// <summary>
+        /// Attempts to get the value of a <see cref="CommandOption"/> given the target <paramref name="type"/>.
+        /// </summary>
+        public static object GetValue(this CommandOption option, Type type)
         {
             switch (option.OptionType)
             {
-                case CommandOptionType.NoValue:
+                case CommandOptionType.NoValue when type == typeof(bool):
                     return option.HasValue();
+
+                case CommandOptionType.SingleValue:
+                case CommandOptionType.SingleOrNoValue:
+                    if (!option.HasValue())
+                        return null;
+
+                    var parsedValue = option.GetType().GetProperty(nameof(CommandOption<object>.ParsedValue));
+                    var value = parsedValue?.GetValue(option) ?? option.Value();
+                    return value != null && type.IsAssignableFrom(value.GetType())
+                        ? value
+                        : null;
+
                 case CommandOptionType.MultipleValue:
-                    return option.Values;
+                    var parsedValues = option.GetType().GetProperty(nameof(CommandOption<object>.ParsedValues));
+                    var values = parsedValues?.GetValue(option) ?? option.Values;
+                    return type.IsAssignableFrom(values.GetType())
+                        ? values
+                        : null;
+
                 default:
-                    return option.Value();
+                    return null;
             }
         }
     }
